fix(android): type mismatch cairkanPinjamanWithFotoOffline — besarPinjaman Int, bukan Long
Build error (PelangganViewModel.kt:5287): "Type mismatch: inferred type is Long
but Int was expected" pada besarPinjaman = besarPinjaman di pemanggilan
offlineRepo.queueSerahTerima.

PENYEBAB: di commit 7b43376 saya deklarasi `val besarPinjaman =
pelanggan.besarPinjaman.toLong()` padahal queueSerahTerima(besarPinjaman: Int)
mengharapkan Int. Pelanggan.besarPinjaman memang sudah Int.

FIX: hapus `.toLong()` — `val besarPinjaman = pelanggan.besarPinjaman` (Int).
Tidak ada konsumen Long lain di fungsi ini. Logic lain utuh, brace seimbang
(baseline PVM +1 pre-existing).

+++ app/src/main/kotlin/com/example/koperasikitagodangulu/PelangganViewModel.kt
@@ -5270,7 +5270,8 @@ class PelangganViewModel(application: Application) : AndroidViewModel(applicatio
                 }
 
                 // 5) Enqueue Storage upload + cleanse + notif via Room → WorkManager.
-                val besarPinjaman = pelanggan.besarPinjaman.toLong()
+                // besarPinjaman dipertahankan Int — sesuai param queueSerahTerima(besarPinjaman: Int).
+                val besarPinjaman = pelanggan.besarPinjaman
                 val tenor = pelanggan.tenor
                 val adminName = try {
                     database.child("metadata/admins/$adminUid/name").get().await()
